+++ packages/flutter_bluetooth_printer/lib/flutter_bluetooth_printer_library.dart
@@ -2,7 +2,6 @@ library flutter_bluetooth_printer;
 
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
 import 'dart:ui';
 
 import 'package:flutter/foundation.dart';