+++ packages/flutter_bluetooth_printer_web/lib/flutter_bluetooth_printer_web.dart
@@ -1,0 +1,1 @@
+export 'flutter_bluetooth_printer_web_library.dart';